moved functions
