+++ src/tokenFactory/jwt/did.ts
@@ -1,0 +1,9 @@
+import { Network, NetworkID } from '../../../typings'
+
+export function createDID(
+  network: Network, // TODO: Is this correct?
+  proxyAddress: string,
+  networkID: NetworkID // TODO: Is this correct?
+) {
+  return `did:ala:${network}:${networkID}:${proxyAddress}`
+}
