+++ src/tokenFactory/jwt/index.ts
@@ -4,3 +4,7 @@ export { createAlastriaToken } from './alastriaToken'
 export { createCredential } from './credential'
 export { createPresentation } from './presentation'
 export { createPresentationRequest } from './presentationRequest'
+
+export { decodeJWT, signJWT, verifyJWT } from './crypto'
+export { createDID } from './did'
+export { PSMHash } from './PSMHash'
